updated code of product and bookRelease
