+++ src/components/ProductInfo.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function Information({
+export default function ProductInfo({
     heading = 'Meal Plan',
     title = 'this is meal',
     author = 'author name',
@@ -10,7 +10,7 @@ export default function Information({
         <h2 className="heading uppercase text-slate-300 ">{heading}</h2>
         <p className=' title hover:underline text-2xl my-5 text-black font-semibold'>{title}</p>
         <div className="divider"></div> 
-        <h3 className=' authorname uppercase text-red-600'>{author}</h3>
+        <h3 className=' author name uppercase text-red-600'>{author}</h3>
     </div>
   )
 }
