+++ src/components/pages/Home/BookRelease.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import allBooksImg from '../../../assets/images/allBooks.png';
-import Information from '../../SectionTitle';
+import ProductInfo from "../../ProductInfo";
 
-export default function Books() {
+export default function BookRelease() {
   return (
     <div>
         <div className="container">
@@ -11,7 +11,7 @@ export default function Books() {
                 <img src={allBooksImg} alt='All the books' />
             </div>
             <div className="information bg-white h-60 w-2/5 -mt-16 relative  ml-96">
-                <Information heading='shopping' title='34 New Baking Books Comming This Fall' author='By Jacqueline Raposo' />
+                <ProductInfo heading='shopping' title='34 New Baking BookRelease Coming This Fall' author='By Jacqueline Raposo' />
             </div>
             </a>
         </div>
