+++ src/components/pages/Home/Home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Books from "./Books";
+import BookRelease from "./BookRelease";
 import HomeCard from "./HomeCard";
 import Products from "./Products";
 
@@ -7,7 +7,7 @@ const Home = () => {
   return (
     <div>
       <HomeCard></HomeCard>
-      <Books></Books>
+      <BookRelease></BookRelease>
       <Products></Products>
     </div>
   );
