+++ src/components/pages/Home/Products.js
@@ -1,14 +1,17 @@
 import React from 'react'
-import Product from './Product'
-
+import Product from '../../Product';
 import moonCakesImg from '../../../assets/images/Mooncakes_HERO_091321_20978.jpg';
 import readPaperPastaImg from '../../../assets/images/RedPepperPasta_HERO_040821_12876.jpg';
 
 export default function Products() {
   return (
-    <div className='products flex justify-center'>
-        <Product heading='meal plan' title='A Week of Easy Dinners to Get You Back in the Cooking Groove' author='by anna hezel' img={moonCakesImg}/>
-        <Product heading='dessert' title='A Week of Easy Dinners to Get You Back in the Cooking Groove' author='By Genevieve Yam' img={readPaperPastaImg}/>
-    </div>
+      <div className='container'>
+          <div className='products flex justify-center'>
+              <a href='/'>
+                <Product heading='meal plan' title='A Week of Easy Dinners to Get You Back in the Cooking Groove' author='by anna hezel' img={moonCakesImg}/>
+                <Product heading='dessert' title='A Week of Easy Dinners to Get You Back in the Cooking Groove' author='By Genevieve Yam' img={readPaperPastaImg}/>
+              </a>
+          </div>
+      </div>
   )
 }
